split DockerPlatformHelper into 2
